+++ src/nx3d/core.py
@@ -1,7 +1,7 @@
 """ This source provides functionality for plotting nodes and edges of nx.Graph objects in 3D.
 """
 
-from math import atan, cos, isclose, pi, sin, sqrt
+from math import cos, isclose, pi, sin, sqrt, tan
 from pathlib import Path
 from typing import Any, Callable, Optional, Union
 
@@ -17,10 +17,12 @@ from panda3d.core import (
     NodePath,
     PointLight,
     TextNode,
+    Vec3,
+    Vec4,
 )
 
 from nx3d import utils
-from nx3d.types import Pos3, Vec3, Vec4
+from nx3d.types import Pos3
 
 FILES = {
     "node": Path(__file__).parent / "data/icosphere.egg",
@@ -44,10 +46,10 @@ UPS = 32  # updates per second to state
 DEFAULTS = dict(
     node_shape=0.9,
     node_color=(0.4, 0, 0.3, 1),
-    node_text_color=(0, 1, 0, 1),
+    node_label_color=(0, 1, 0, 1),
     edge_radius=1.75,
     edge_color=(0.3, 0.3, 0.3, 0.5),
-    edge_text_color=(0, 1, 0, 1),
+    edge_label_color=(0, 1, 0, 1),
     speed_theta=96.0,
     speed_phi=96.0,
     speed_radius=36.0,
@@ -89,6 +91,7 @@ class Nx3D(ShowBase):
             Set attributes on graph components to update the render. If not None, the graph's nodes and edges must be
             annotated with 'color' and 'label' entries in the annotation dictionary i.e. g.nodes[nd]['color'] must exist for
             all nodes.
+        kwagrs: Passed to super().__init__
 
     Returns:
         ShowBase: The Panda3D object capable of rendering the graph
@@ -101,22 +104,24 @@ class Nx3D(ShowBase):
         node_color: Vec4 = DEFAULTS["node_color"],
         node_shape: Union[float, Vec3] = DEFAULTS["node_shape"],
         node_labels: dict = {},
-        node_label_color: Vec4 = DEFAULTS["node_text_color"],
+        node_label_color: Vec4 = DEFAULTS["node_label_color"],
         edge_color: Vec4 = DEFAULTS["edge_color"],
         edge_radius: Union[float, list[float]] = DEFAULTS["edge_radius"],
         edge_labels: dict = {},
-        edge_label_color: Vec4 = DEFAULTS["edge_text_color"],
+        edge_label_color: Vec4 = DEFAULTS["edge_label_color"],
         plot_axes=False,
         verbose=False,
         autolabel=False,
         mouse=False,
         state_trans_freq: float = 1.0,
         state_trans_func: Optional[Callable[[nx.Graph, int, float], Any]] = None,
+        **kwargs,
     ):
-        ShowBase.__init__(self)
+        ShowBase.__init__(self, **kwargs)
         self.g = graph
         self.verbose = verbose
 
+        # init file paths
         node_fp = FILES.get("node")
         if isinstance(self.g, nx.MultiDiGraph):
             edge_fp = FILES.get("edge_directed_bent")
@@ -129,19 +134,44 @@ class Nx3D(ShowBase):
         if edge_fp is None or node_fp is None:
             raise NotImplementedError
 
-        if pos is None:
-            pos_scale = 2.0 * sqrt(len(self.g.nodes))
-            pos = nx.spring_layout(self.g, dim=3, scale=pos_scale)
-        if not all(len(p) == 3 for p in pos.values()):
-            raise ValueError("pos must be 3d, use the dim=3 kwarg in nx layouts")
-        self.pos = pos
+        # init labels
         if autolabel:
             if verbose and any([edge_labels, node_labels]):
                 print("overwriting labels, set autolabel False if undesired")
             node_labels = {nd: str(nd) for nd in self.g.nodes}
             edge_labels = {ed: str(ed) for ed in self.g.edges}
 
-        # add some lights
+        # init positions
+        if pos is None:
+            pos_scale = 2.0 * sqrt(len(self.g.nodes))
+            pos = nx.spring_layout(self.g, dim=3, scale=pos_scale)
+        if not all(len(p) == 3 for p in pos.values()):
+            raise ValueError("pos must be 3d, use the dim=3 kwarg in nx layouts")
+
+        # init graph attributes
+        for nd in self.g.nodes:
+            self.g.nodes[nd]["pos"] = self.g.nodes[nd].get("pos", pos[nd])
+            self.g.nodes[nd]["color"] = self.g.nodes[nd].get("color", node_color)
+            self.g.nodes[nd]["shape"] = self.g.nodes[nd].get("shape", node_color)
+            self.g.nodes[nd]["label"] = self.g.nodes[nd].get(
+                "label", node_labels.get(nd)
+            )
+            self.g.nodes[nd]["label_color"] = self.g.nodes[nd].get(
+                "label_color", node_label_color
+            )
+        for ed in self.g.edges:
+            self.g.edges[ed]["color"] = self.g.edges[ed].get("color", edge_color)
+            self.g.edges[ed]["radius"] = self.g.edges[ed].get("radius", edge_color)
+            self.g.edges[ed]["label"] = self.g.edges[ed].get(
+                "label", edge_labels.get(nd)
+            )
+            self.g.edges[ed]["label_color"] = self.g.edges[ed].get(
+                "label_color", edge_label_color
+            )
+
+        self.initial_pos = {nd: self.g.nodes[nd]["pos"] for nd in self.g.nodes}
+
+        # init lights
         for i, dl in enumerate(DEFAULTS["light_direct"]):
             hpr = dl["hpr"]
             dlight = DirectionalLight(f"directional_light_{i}")
@@ -161,16 +191,18 @@ class Nx3D(ShowBase):
             self.render.setLight(plnp)
             plnp.setPos(*pos)
 
+        # init 3d models
         for i, nd in enumerate(self.g.nodes):
             pid = f"node_{i}"
             model: NodePath = self._init_panda3d_model(pid, node_fp)
             model.reparentTo(self.render)
             model.setScale(node_shape)
-            utils.set_color(model, node_color)
-            model.setPos(*self.pos[nd])
+            utils.set_color(model, self.g.nodes[nd]["color"])
+            model.setPos(*self.g.nodes[nd]["pos"])
             tpid = f"node_{i}_text"
-            label = node_labels.get(nd)
-            text, text_ = self._init_panda3d_text(tpid, label, node_label_color)
+            text, text_ = self._init_panda3d_text(
+                tpid, self.g.nodes[nd]["label"], self.g.nodes[nd]["label_color"]
+            )
             text.reparentTo(model)
             text.setScale(tuple(1 / sc for sc in model.getScale()))
             text.setZ(model.getBounds().getRadius() * 1.1)
@@ -182,10 +214,11 @@ class Nx3D(ShowBase):
             pid = f"edge_{i}"
             model: NodePath = self._init_panda3d_model(pid, edge_fp)
             model.reparentTo(self.render)
-            utils.set_color(model, edge_color)
+            utils.set_color(model, self.g.edges[ed]["color"])
             # TODO use model.lookAt(node)
             # rotate into place
-            p0, p1 = (self.pos[e] for e in ed)
+            p0 = self.g.nodes[ed[0]]["pos"]
+            p1 = self.g.nodes[ed[1]]["pos"]
             model.setPos(*p0)
             dist = sqrt(((p0 - p1) ** 2).sum())
             model.setScale(edge_radius, edge_radius, dist)
@@ -205,8 +238,9 @@ class Nx3D(ShowBase):
             model.setH(heading)
 
             tpid = f"edge_{i}_text"
-            label = edge_labels.get(ed)
-            text, text_ = self._init_panda3d_text(tpid, label, edge_label_color)
+            text, text_ = self._init_panda3d_text(
+                tpid, self.g.edges[ed]["label"], self.g.edges[ed]["label_color"]
+            )
             text.reparentTo(self.render)
             text.setPos(tuple((p0 + p1) / 2.0))
             self.g.edges[ed]["model"] = model
@@ -226,8 +260,6 @@ class Nx3D(ShowBase):
         state_trans_func,
     ):
         self.state_trans_func = state_trans_func
-        if self.verbose:
-            print(f"state_trans_func {self.state_trans_func}")
         if self.state_trans_func:
             self.taskMgr.doMethodLater(
                 state_trans_freq, self.stateUpdateTask, "StateUpdate"
@@ -279,13 +311,14 @@ class Nx3D(ShowBase):
         self.disableMouse()
         bd = self.render.getBounds()
         rad = np.linalg.norm(bd.getApproxCenter()) + bd.getRadius()
-        fov = min(self.camLens.fov)  # angle in degrees
+        fov = min(self.camLens.fov) if self.camLens else 45
         radians_fov = fov / 180 * pi
-        self.initial_camera_radius = rad / atan(radians_fov) * 1.8
-        self.camera.setPos(0, -self.initial_camera_radius, 0)
+        self.initial_camera_radius = rad / tan(radians_fov) * 1.75
         self.cam_radius = self.initial_camera_radius
         self.cam_theta = 0.0
         self.cam_phi = 0.0
+        if self.camera:
+            self.camera.setPos(0, -self.initial_camera_radius, 0)
         self.taskMgr.add(self.keyboardCameraTask, "KeyboardCameraTask")
 
     def _init_axes(self, fn):
@@ -326,7 +359,7 @@ class Nx3D(ShowBase):
         """main state update loop"""
         if self.verbose:
             print(f"stateUpdateTask from {task.name}")
-        self.state_trans_func(self.g, task.frame, task.time)
+        self.state_trans_func(self.g, task.frame, task.delayTime)
         for elm in utils.all_elements(self.g):
             assert all(isinstance(x, NodePath) for x in (elm["model"], elm["text_np"]))
             assert isinstance(elm["text_tn"], TextNode)
@@ -339,6 +372,8 @@ class Nx3D(ShowBase):
 
     def guiUpdateTask(self, task):
         """write diagnostic info to gui"""
+        if self.camera is None:
+            return
         rot = "camera rotation: "
         pos = "camera position: "
         rot_ = rot + str(self.camera.getHpr())