+++ src/nx3d/types.py
@@ -1,5 +1,5 @@
 from typing import Hashable
 
-Vec3 = tuple[float, float, float]
-Vec4 = tuple[float, float, float, float]
+from panda3d.core import Vec3
+
 Pos3 = dict[Hashable, Vec3]