+++ tests/conftest.py
@@ -1,0 +1,7 @@
+import networkx as nx
+from pytest import fixture
+
+
+@fixture
+def g():
+    return nx.frucht_graph()