+++ tests/test_nx3d.py
@@ -1,0 +1,9 @@
+import nx3d
+
+
+def test_import():
+    ...
+
+
+def test_instantiation(g):
+    _ = nx3d.Nx3D(g, windowType="none")