+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Sun Sep 19 2022
+22.9.9
+- no longer deepcopying the graph passed to the plotting function
+- using nx graph attributes to track desired 3d model state
+- added CI and tests
+
 22.9.8
 - implemented edge attribute interface for 'label' and 'color'